Refactor block device using polly.
Merge BlockEpollHandler with Block and implement EventHandler
for Block. Add a dummy implementation of EventHandler for Net and Vsock.

Adds EventManager in VmmController and uses it Builder to register the
Block device.

Adds a temporary second path to register devices in MmioDeviceManager
with register_block_device() and a generic device resources registration
fn: register_device_resources(). We will remove these once we have all
devices migrated to EventManager.

Signed-off-by: Andrei Sandu <sandreim@amazon.com>